ci: add stale bot configuration file
A stale.yml configuration file was added to control behavior of stale bot. After 60 days of inactivity, an issue will be marked as stale with the label 'wontfix'. If a further 7 days pass without activity, the issue will be closed. Issues labelled as 'pinned' or 'security' are exempt from becoming stale. This is to automate the management of inactive issues and keep the issue tracker tidy.