feat: add GitHub action to greet user and log time
A new 'Hello World' GitHub action has been added. The action greets a person specified by the 'who-to-greet' input. The greeting is logged as well as the current UTC time. 'entrypoint.sh' implements the greeting and logging functions, while 'action.yml' specifies the action's inputs and outputs. This commit aims to create a more interactive user experience and allows for customizable greetings within the workflow.