fix: update syntax in time output in `entrypoint.sh`
The previous usage of brackets in the echo statement of the entrypoint.sh was causing issues in the correct interpretation of the time output. We have replaced the brackets with "=" to fix an issue with the output of date and time in entrypoint.sh. This change will ensure that the date and time format is saved correctly in GITHUB_ENV, improving the reliability and readability of logs.