fix: update output value and additional parameters in `action.yml`
Modified the 'action.yml' to include the 'steps.greet.outputs.time' as the value for the 'time' output. Introduced 'name' and 'id' parameters to 'run' object to allow the usage of output from this step in other steps. Also, linked the 'INPUT_WHO_TO_GREET' environmental variable directly to 'who-to-greet' input for more straightforward reference.